Remove the interface
