+++ lucene/core/src/java/org/apache/lucene/search/QueueSizeBasedExecutionControlPlane.java
@@ -25,13 +25,14 @@ import java.util.concurrent.FutureTask;
 import java.util.concurrent.ThreadPoolExecutor;
 
 /**
- * Implementation of SliceBasedExecutionControlPlane that controls the number of active threads
+ * Derivative of SliceBasedExecutionControlPlane that controls the number of active threads
  * that are used for a single query. At any point, no more than (maximum pool size of the executor * LIMITING_FACTOR)
  * threads should be active. If the limit is exceeded, further segments are searched on the caller thread
  */
-public class QueueSizeBasedExecutionControlPlane extends DefaultSliceExecutionControlPlane {
-  private final ThreadPoolExecutor threadPoolExecutor;
+public class QueueSizeBasedExecutionControlPlane extends SliceExecutionControlPlane {
   private static final double LIMITING_FACTOR = 1.5;
+
+  private final ThreadPoolExecutor threadPoolExecutor;
 
   public QueueSizeBasedExecutionControlPlane(ThreadPoolExecutor threadPoolExecutor) {
     super(threadPoolExecutor);
